[ao] Created Skeleton for ModelReportVisualization class
Summary: This introduces the skeleton for the ModelReportVisualization
class. This class helps visualize the information generated by the
ModelReport class `generate_report()` output. This class aims to provide
visualizations in a table, plot (line graph) and histogram view.

This also introduces an empty test class for testing visualizations. As
implementations start occuring for this class, tests will also be
approrpriately added.

This includes the high level descriptions for each of the methods as
well. Expected use cases will be added to the class description in a
future commit as that gets finalized.

Test Plan: python test/test_quantization.py TestFxModelReportVisualization

Reviewers:

Subscribers:

Tasks:

Tags:

ghstack-source-id: 37a406b859f9bc2aaec2427b8cc78c3fd1486cb9
Pull Request resolved: https://github.com/pytorch/pytorch/pull/81523